project/requirements.txt
PyQt5==5.15.9
PyYAML==6.0.1
setuptools>=65.5.1